Added an interface and usage instructions.
